Make RTS compilable for enhanced orthogonal persistence

+++ rts/motoko-rts/src/idl.rs
@@ -12,11 +12,7 @@ use core::cmp::min;
 
 use motoko_rts_macros::{classical_persistence, enhanced_orthogonal_persistence, ic_mem_fn};
 
-#[enhanced_orthogonal_persistence]
 use crate::libc_declarations::{c_void, memcmp};
-
-#[classical_persistence]
-use libc::{c_void, memcmp};
 
 //
 // IDL constants
@@ -1245,6 +1241,8 @@ unsafe fn idl_sub<M: Memory>(
     t1: i32,
     t2: i32,
 ) -> bool {
+    use crate::persistence::compatibility::TypeDescriptor;
+
     let mut type_descriptor2 = TypeDescriptor::new(candid_data2, type_offsets2);
     let typtbl2 = type_descriptor2.build_type_table(mem);
     let typtbl_end2 = type_descriptor2.type_table_end();
@@ -1270,13 +1268,13 @@ unsafe extern "C" fn idl_sub(
 }
 
 unsafe fn idl_sub_internal(
-    rel_buf: *mut u32, // a buffer with at least 2 * typtbl_size1 * typtbl_size2 bits
+    rel_buf: *mut usize, // a buffer with at least 2 * typtbl_size1 * typtbl_size2 bits
     typtbl1: *mut *mut u8,
     typtbl2: *mut *mut u8,
     typtbl_end1: *mut u8,
     typtbl_end2: *mut u8,
-    typtbl_size1: u32,
-    typtbl_size2: u32,
+    typtbl_size1: usize,
+    typtbl_size2: usize,
     t1: i32,
     t2: i32,
 ) -> bool {
